requirements.txt
Django==3.0.2
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
psycopg2==2.8.4
pylint==2.4.4
pytest==5.3.4
